feat: 모킹 구조 변경 및 points, cheer 핸들러 추가
관리하기 쉽도록 분리했습니다.
