+++ mocks/apiHandlers/db.ts
@@ -1,0 +1,191 @@
+const myMainDB = {
+  goals: [
+    {
+      id: "골1",
+      title: "자격증 따기",
+      dueDate: "2025-07-01",
+      progress: 24.5,
+      isCompleted: true,
+      isJoinedGroup: true,
+      subGoals: ["서브골1", "서브골2"],
+    },
+    {
+      id: "골2",
+      title: "프로젝트 완주하기",
+      dueDate: "2025-09-10",
+      progress: 50,
+      isCompleted: false,
+      isJoinedGroup: true,
+      subGoals: ["서브골3", "서브골4"],
+    },
+  ],
+  subGoals: [
+    {
+      id: "서브골1",
+      title: "자격증 책 한 권 끝내기",
+      todos: ["투두1", "투두2"],
+      isCompleted: false,
+    },
+    {
+      id: "서브골2",
+      title: "자격증 시험 보기",
+      todos: ["투두3", "투두4"],
+      isCompleted: false,
+    },
+    {
+      id: "서브골3",
+      title: "기획",
+      todos: ["투두5", "투두6", "투두9", "투두10"], // 투두 9, 10 추가
+      isCompleted: false,
+    },
+    {
+      id: "서브골4",
+      title: "개발",
+      todos: ["투두7", "투두8", "투두11", "투두12"], // 투두 11, 12 추가
+      isCompleted: false,
+    },
+  ],
+  todos: [
+    {
+      id: "투두1",
+      title: "영단어 10개 이상 외우기",
+      date: "2025-07-01",
+      status: "COMPLETE",
+      todoResultId: "투두결과1",
+      createdAt: "2025-07-01T07:57:53.097Z",
+    },
+    {
+      id: "투두2",
+      title: "암기 테스트 보기",
+      date: "2025-07-03",
+      status: "INCOMPLETE",
+      todoResultId: undefined,
+      createdAt: "2025-07-02T07:57:53.097Z",
+    },
+    {
+      id: "투두3",
+      title: "모의고사 치르기",
+      date: "2025-07-06",
+      status: "INCOMPLETE",
+      todoResultId: undefined,
+      createdAt: "2025-07-02T07:57:53.097Z",
+    },
+    {
+      id: "투두4",
+      title: "오답 분석하기",
+      date: "2025-07-08",
+      status: "INCOMPLETE",
+      todoResultId: undefined,
+      createdAt: "2025-07-02T07:57:53.097Z",
+    },
+    {
+      id: "투두5",
+      title: "프로젝트 기획서 초안 작성",
+      date: "2025-07-05",
+      status: "COMPLETE",
+      todoResultId: "투두결과2",
+      createdAt: "2025-07-02T07:57:53.097Z",
+    },
+    {
+      id: "투두6",
+      title: "팀원들과 기획 회의",
+      date: "2025-07-07",
+      status: "INCOMPLETE",
+      todoResultId: undefined,
+      createdAt: "2025-07-02T07:57:53.097Z",
+    },
+    {
+      id: "투두7",
+      title: "개발 환경 세팅",
+      date: "2025-07-09",
+      status: "COMPLETE",
+      todoResultId: "투두결과3",
+      createdAt: "2025-07-02T07:57:53.097Z",
+    },
+    {
+      id: "투두8",
+      title: "백엔드 API 설계",
+      date: "2025-07-11",
+      status: "INCOMPLETE",
+      todoResultId: undefined,
+      createdAt: "2025-07-02T07:57:53.097Z",
+    },
+    {
+      id: "투두9",
+      title: "사용자 스토리 정의",
+      date: "2025-07-06",
+      status: "COMPLETE",
+      todoResultId: "투두결과4",
+      createdAt: "2025-07-03T07:57:53.097Z",
+    },
+    {
+      id: "투두10",
+      title: "UI/UX 와이어프레임 제작",
+      date: "2025-07-08",
+      status: "INCOMPLETE",
+      todoResultId: undefined,
+      createdAt: "2025-07-03T07:57:53.097Z",
+    },
+    {
+      id: "투두11",
+      title: "프론트엔드 컴포넌트 개발",
+      date: "2025-07-12",
+      status: "INCOMPLETE",
+      todoResultId: undefined,
+      createdAt: "2025-07-03T07:57:53.097Z",
+    },
+    {
+      id: "투두12",
+      title:
+        "데이터베이스 스키마 설계 겁나 길게 쓸거임 일부러 어디까지 가나 보자고 아니 좀 더 길어야 하는 것 같은뎅 와 어디까지 길어져야만 하는 겁니까.",
+      date: "2025-07-10",
+      status: "COMPLETE",
+      todoResultId: "투두결과5",
+      createdAt: "2025-07-03T07:57:53.097Z",
+    },
+  ],
+  todoResults: [
+    {
+      todoResultId: "투두결과1",
+      todoId: "투두1",
+      emotion: "뿌듯",
+      content: "영단어 10개 이상 외우기를 했다.",
+      fileUrl: "url 대충",
+    },
+    {
+      todoResultId: "투두결과2",
+      todoId: "투두5",
+      emotion: "만족",
+      content: "프로젝트 기획서 초안을 성공적으로 작성했다.",
+      fileUrl: "기획서_초안_링크",
+    },
+    {
+      todoResultId: "투두결과3",
+      todoId: "투두7",
+      emotion: "기대",
+      content: "개발 환경 세팅 완료! 이제 본격적인 개발 시작이다.",
+      fileUrl: "개발환경_세팅_스크린샷",
+    },
+    {
+      todoResultId: "투두결과4",
+      todoId: "투두9",
+      emotion: "흥미로움",
+      content: "사용자 스토리 정의를 통해 기능 구현 방향성을 잡았다.",
+      fileUrl: "사용자_스토리_문서",
+    },
+    {
+      todoResultId: "투두결과5",
+      todoId: "투두12",
+      emotion: "개운함",
+      content: "데이터베이스 스키마 설계를 마치고, 데이터 모델링을 완료했다.",
+      fileUrl: "DB_스키마_다이어그램",
+    },
+  ],
+  cheer: {
+    default: "목표는 멀어도 나는 계속 가는중",
+  },
+  points: {
+    value: 1000,
+  },
+};
+export { myMainDB };
