+++ mocks/apiHandlers/utils.ts
@@ -1,0 +1,6 @@
+import { HttpResponse } from "msw";
+
+const createErrorResponse = () =>
+  new HttpResponse("네트워크 에러", { status: 500 });
+
+export { createErrorResponse };
